backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
bcrypt==4.2.1
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.115.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.0
httptools==0.7.1
idna==3.11
openai==1.59.8
langchain==0.3.26
langchain-community==0.3.27
sentence-transformers==5.1.0
passlib==1.7.4
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pyparsing==3.2.5
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
faiss-cpu==1.9.0.post1
pypdf==5.1.0
tiktoken==0.8.0
numpy==2.1.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.31
starlette==0.38.6
tqdm==4.67.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.0
uvicorn==0.30.0
# uvloop==0.22.1  # Linux/Mac only, not needed for Railway
watchfiles==1.1.1
websockets==15.0.1
